Rearrange classes to have a features specific class
