+++ games_services/lib/src/player.dart
@@ -1,0 +1,34 @@
+import 'package:games_services_platform_interface/game_services_platform_interface.dart';
+import 'package:games_services_platform_interface/models.dart';
+
+abstract class Player {
+  /// Show the iOS Access Point.
+  static Future<String?> showAccessPoint(AccessPointLocation location) async {
+    return await GamesServicesPlatform.instance.showAccessPoint(location);
+  }
+
+  /// Hide the iOS Access Point.
+  static Future<String?> hideAccessPoint() async {
+    return await GamesServicesPlatform.instance.hideAccessPoint();
+  }
+
+  /// Get the player id.
+  /// On iOS the player id is unique for your game but not other games.
+  static Future<String?> getPlayerID() async {
+    return await GamesServicesPlatform.instance.getPlayerID();
+  }
+
+  /// Get the player name.
+  /// On iOS the player alias is the name used by the Player visible in the leaderboard
+  static Future<String?> getPlayerName() async {
+    return await GamesServicesPlatform.instance.getPlayerName();
+  }
+
+  /// Get player score for a specific leaderboard.
+  static Future<int?> getPlayerScore(
+      {iOSLeaderboardID = "", androidLeaderboardID = ""}) async {
+    return await GamesServicesPlatform.instance.getPlayerScore(
+        iOSLeaderboardID: iOSLeaderboardID,
+        androidLeaderboardID: androidLeaderboardID);
+  }
+}
