+++ games_services_platform_interface/lib/models.dart
@@ -1,0 +1,5 @@
+export 'src/models/access_point_location.dart';
+export 'src/models/achievement.dart';
+export 'src/models/leaderboard_scope.dart';
+export 'src/models/leaderboard_time_scope.dart';
+export 'src/models/score.dart';
