+++ games_services_platform_interface/lib/src/models/achievement.dart
@@ -1,0 +1,18 @@
+import '../util/device.dart';
+
+class Achievement {
+  String? androidID;
+  String iOSID;
+  double percentComplete;
+  int steps;
+
+  String? get id {
+    return Device.isPlatformAndroid ? androidID : iOSID;
+  }
+
+  Achievement(
+      {this.androidID,
+      this.iOSID = "",
+      this.percentComplete = 100,
+      this.steps = 0});
+}
