+++ games_services_platform_interface/lib/src/util/device.dart
@@ -1,5 +1,5 @@
 import 'dart:io' show Platform;
 
-class Helpers {
+class Device {
   static var isPlatformAndroid = Platform.isAndroid;
 }
